# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Time Off Service API
  version: 0.1.0
  description: |
    MyCompany time-off microservice. JSON:API v1.1 over `application/vnd.api+json`.

    **Authentication:** Bearer JWT (HS256). Mint test tokens with `npm run tools:jwt`.

    **Correlation:** Send optional `X-Correlation-Id`; echoed in responses and error `meta`.

    Documents **implemented** Phase 1 and Phase 2 endpoints. See [spec.md](spec.md) for the full roadmap.
  contact:
    name: Time Off Service
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/leave-requests:
    get:
      tags: [Leave Requests]
      summary: List leave requests
      description: |
        Paginated list. Non-privileged users see only their own requests.
        Privileged roles may filter by `filter[employeeId]` and `filter[status]`.
      operationId: listLeaveRequests
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
        - name: filter[status]
          in: query
          schema:
            type: string
            enum:
              - DRAFT
              - PENDING
              - APPROVED_PENDING_HCM_UPDATE
              - APPROVED
              - REJECTED
              - CANCELLED
        - name: filter[employeeId]
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Leave request collection
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/LeaveRequestCollectionDocument'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Leave Requests]
      summary: Create leave request
      description: "Creates a leave request. Set submit: true to submit for approval immediately."
      operationId: createLeaveRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              $ref: '#/components/schemas/CreateLeaveRequestBody'
            example:
              data:
                type: leave-requests
                attributes:
                  startDate: '2026-07-10'
                  endDate: '2026-07-12'
                  partialDay: NONE
                  reason: Family travel
                  submit: true
                  dimensions:
                    locationId: US-NY
                relationships:
                  employee:
                    data:
                      type: employees
                      id: '{employeeId}'
                  leaveType:
                    data:
                      type: leave-types
                      id: '{leaveTypeId}'
      responses:
        '201':
          description: Created leave request
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/LeaveRequestDocument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/leave-requests/{id}:
    get:
      tags: [Leave Requests]
      summary: Get leave request
      operationId: getLeaveRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LeaveRequestId'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Leave request resource
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/LeaveRequestDocument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/leave-requests/{id}/cancel:
    post:
      tags: [Leave Requests]
      summary: Cancel leave request
      description: Cancels a request in `PENDING`, `APPROVED_PENDING_HCM_UPDATE`, `APPROVED`, or `DRAFT` status.
      operationId: cancelLeaveRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LeaveRequestId'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Cancelled leave request
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/LeaveRequestDocument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/leave-requests/{id}/approve:
    post:
      tags: [Approvals]
      summary: Approve leave request
      description: Approves the current workflow step. Posts to HCM when fully approved.
      operationId: approveLeaveRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LeaveRequestId'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/vnd.api+json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionBody'
      responses:
        '200':
          description: Updated leave request
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/LeaveRequestDocument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/leave-requests/{id}/reject:
    post:
      tags: [Approvals]
      summary: Reject leave request
      operationId: rejectLeaveRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LeaveRequestId'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/vnd.api+json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionBody'
      responses:
        '200':
          description: Rejected leave request
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/LeaveRequestDocument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/approvals/pending:
    get:
      tags: [Approvals]
      summary: List pending approvals
      description: Pending approval steps for the authenticated manager. Requires `manager`, `hr_admin`, or `system_admin`.
      operationId: listPendingApprovals
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Approval collection
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ApprovalCollectionDocument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT. Claims: `sub`, `roles[]`, optional `employeeId`.
        Roles: `employee`, `manager`, `hr_admin`, `system_admin`, `integration_client`.
  parameters:
...
    LeaveRequestId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageNumber:
      name: page[number]
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page[size]
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
      description: Optional trace ID; generated if omitted
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Optional idempotency key for safe POST retries (24h TTL)
  headers:
    CorrelationId:
      description: Request correlation ID
      schema:
        type: string
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/vnd.api+json:
          schema:
            $ref: '#/components/schemas/JsonApiErrorDocument'
    Forbidden:
      description: Insufficient permissions
      content:
        application/vnd.api+json:
          schema:
            $ref: '#/components/schemas/JsonApiErrorDocument'
    NotFound:
      description: Resource not found
      content:
        application/vnd.api+json:
          schema:
            $ref: '#/components/schemas/JsonApiErrorDocument'
    UnprocessableEntity:
      description: Business rule or validation failure
      content:
        application/vnd.api+json:
          schema:
            $ref: '#/components/schemas/JsonApiErrorDocument'
  schemas:
...
    JsonApiVersion:
      type: object
      required: [version]
      properties:
        version:
          type: string
          example: '1.1'
    JsonApiError:
      type: object
      properties:
        status:
          type: string
        code:
          type: string
        title:
          type: string
        detail:
          type: string
        source:
          type: object
          properties:
            pointer:
              type: string
    JsonApiErrorDocument:
      type: object
      required: [jsonapi, errors]
      properties:
        jsonapi:
          $ref: '#/components/schemas/JsonApiVersion'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/JsonApiError'
        meta:
          type: object
          properties:
            correlationId:
              type: string
    PaginationMeta:
      type: object
      properties:
        totalCount:
          type: integer
        pageNumber:
          type: integer
        pageSize:
          type: integer
    PaginationLinks:
      type: object
      properties:
        self:
          type: string
        next:
          type: string
        prev:
          type: string
    ResourceIdentifier:
      type: object
      required: [type, id]
      properties:
        type:
          type: string
        id:
          type: string
    RelationshipData:
      type: object
      properties:
        data:
          oneOf:
            - $ref: '#/components/schemas/ResourceIdentifier'
            - type: 'null'
...
    LeaveRequestAttributes:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        durationDays:
          type: number
        partialDayType:
          type: string
          enum: [NONE, START, END, BOTH]
        partialDayHours:
          type: number
        dimensions:
          type: object
          additionalProperties: true
        status:
          type: string
          enum:
            - DRAFT
            - PENDING
            - APPROVED_PENDING_HCM_UPDATE
            - APPROVED
            - REJECTED
            - CANCELLED
        reason:
          type: string
        submittedAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
        hcmReferenceId:
          type: string
        hcmPostedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LeaveRequestResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [leave-requests]
        id:
          type: string
          format: uuid
        attributes:
          $ref: '#/components/schemas/LeaveRequestAttributes'
        relationships:
          type: object
          properties:
            employee:
              $ref: '#/components/schemas/RelationshipData'
            leaveType:
              $ref: '#/components/schemas/RelationshipData'
    LeaveRequestDocument:
      type: object
      required: [jsonapi, data]
      properties:
        jsonapi:
          $ref: '#/components/schemas/JsonApiVersion'
        data:
          $ref: '#/components/schemas/LeaveRequestResource'
    LeaveRequestCollectionDocument:
      type: object
      required: [jsonapi, data]
      properties:
        jsonapi:
          $ref: '#/components/schemas/JsonApiVersion'
        data:
          type: array
          items:
            $ref: '#/components/schemas/LeaveRequestResource'
        links:
          $ref: '#/components/schemas/PaginationLinks'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    CreateLeaveRequestBody:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [type, attributes, relationships]
          properties:
            type:
              type: string
              enum: [leave-requests]
            attributes:
              type: object
              required: [startDate, endDate]
              properties:
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                partialDay:
                  type: string
                  enum: [NONE, START, END, BOTH]
                  default: NONE
                partialDayHours:
                  type: number
                reason:
                  type: string
                submit:
                  type: boolean
                  default: false
                documentationProvided:
                  type: boolean
                dimensions:
                  type: object
                  additionalProperties: true
            relationships:
              type: object
              required: [leaveType]
              properties:
                employee:
                  type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/ResourceIdentifier'
                leaveType:
                  type: object
                  required: [data]
                  properties:
                    data:
                      $ref: '#/components/schemas/ResourceIdentifier'
    ApprovalDecisionBody:
      type: object
      properties:
        data:
          type: object
          properties:
            type:
              type: string
              enum: [approvals]
            attributes:
              type: object
              properties:
                comment:
                  type: string
    ApprovalAttributes:
      type: object
      properties:
        approvalLevel:
          type: integer
        decision:
          type: string
        comment:
          type: string
        decidedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ApprovalResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [approvals]
        id:
          type: string
          format: uuid
        attributes:
          $ref: '#/components/schemas/ApprovalAttributes'
        relationships:
          type: object
          properties:
            leaveRequest:
              $ref: '#/components/schemas/RelationshipData'
            approver:
              $ref: '#/components/schemas/RelationshipData'
    ApprovalCollectionDocument:
      type: object
      required: [jsonapi, data]
      properties:
        jsonapi:
          $ref: '#/components/schemas/JsonApiVersion'
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApprovalResource'
        links:
          $ref: '#/components/schemas/PaginationLinks'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...